requirements.txt
chromadb==1.3.5
Flask==3.1.2
flask_cors==6.0.1
openai==2.8.1
pandas==2.3.3
python-dotenv==1.2.1
Requests==2.32.5
